+++ src/Public/Get-Dependency.ps1
@@ -1,0 +1,13 @@
+function Get-Dependency {
+
+    [CmdletBinding()]
+    param (
+        [Parameter( Mandatory, ValueFromPipeline )]
+        [ValidateNotNullOrEmpty()]
+        [PsCustomObject] $Activity
+    )
+
+    process {
+        Write-Output $Activity.dependsOn
+    }
+}